refactor: create `scsi.rs`
